Fixing project folder structure
