+++ app/views/users/login.ejs
@@ -42,9 +42,11 @@
                     <br>
                     <div class="row justify-content-center mt-3">
                         <div class="col-11 center">
-                            <button type="submit" class="boton-secundario boton-flex">
-                                Crear Cuenta
-                            </button>
+                            <a href="/user/register">
+                                <button type="button" class="boton-secundario boton-flex">
+                                    Crear Cuenta
+                                </button>
+                            </a>
                         </div>
                     </div>
                     <br><br><br><br><br>
